PuppyList move to PuppyListScreen inside
